+++ src/nodes/crosshair/base.ts
@@ -1,4 +1,5 @@
 import { devGroup } from "$lib/utils";
+import { requestNamedLocation } from "$lib/namedLocations";
 import { TriggerEngine as T } from "trigger-engine/types";
 
 const { TriggerNode } = globalThis.triggerEngine;
@@ -60,7 +61,7 @@ abstract class CrosshairModifierNode<
 
 	protected abstract apply(section: CrosshairSection): Promise<void> | void;
 
-	getLocation(loc: PositionSource): TokenDocument | Point | RegionDocument | undefined {
+	getLocation(loc: PositionSource): TokenDocument | Point | RegionDocument | string | undefined {
 		switch (loc.kind) {
 			case "point":
 				return { x: loc.x, y: loc.y };
@@ -68,6 +69,9 @@ abstract class CrosshairModifierNode<
 				return loc.region;
 			case "target":
 				return this.getTargetToken({ actor: loc.actor, token: loc.token });
+			case "name":
+				requestNamedLocation(this, loc.name);
+				return loc.name;
 		}
 	}