+++ src/nodes/sound/base.ts
@@ -1,4 +1,5 @@
 import { devGroup } from "$lib/utils";
+import { requestNamedLocation } from "$lib/namedLocations";
 import { TriggerEngine as T } from "trigger-engine/types";
 import { EASE_OPTIONS } from "../effect/constants";
 
@@ -71,7 +72,7 @@ abstract class SoundModifierNode<
 		return null;
 	}
 
-	getLocation(loc: PositionSource): TokenDocument | Point | RegionDocument | undefined {
+	getLocation(loc: PositionSource): TokenDocument | Point | RegionDocument | string | undefined {
 		switch (loc.kind) {
 			case "point":
 				return { x: loc.x, y: loc.y };
@@ -79,6 +80,9 @@ abstract class SoundModifierNode<
 				return loc.region;
 			case "target":
 				return this.getTargetToken({ actor: loc.actor, token: loc.token });
+			case "name":
+				requestNamedLocation(this, loc.name);
+				return loc.name;
 		}
 	}